+++ CHANGELOG.md
@@ -31,6 +31,7 @@ corresponding version of the Dune core modules.
   simplify writing of new pre-bases.
 - Add a dynamic power basis with a runtime exponent that can be constructed with the factory
   `power(<sub-basis>, k)` or `power(<sub-basis>, k, <index-merging>)`.
+- Add mixin class `LFEPreBasisMixin` to create leaf pre-bases based on a local finite-element and a layout.
 
 Deprecations and removals